refact: move navigator-related structures to app.navigation.navigator
