+++ app/src/main/java/com/jeanbarrossilva/mastodonte/app/navigation/navigator/BottomNavigationItemNavigatorFactory.kt
@@ -1,6 +1,6 @@
-package com.jeanbarrossilva.mastodonte.app.navigation
+package com.jeanbarrossilva.mastodonte.app.navigation.navigator
 
-import com.jeanbarrossilva.mastodonte.app.navigation.provider.ProfileBottomNavigationItemNavigator
+import com.jeanbarrossilva.mastodonte.app.navigation.navigator.provider.ProfileBottomNavigationItemNavigator
 
 internal object BottomNavigationItemNavigatorFactory {
     fun create(): BottomNavigationItemNavigator {
