+++ app/src/main/java/com/jeanbarrossilva/mastodonte/app/navigation/navigator/provider/ProfileBottomNavigationItemNavigator.kt
@@ -1,8 +1,8 @@
-package com.jeanbarrossilva.mastodonte.app.navigation.provider
+package com.jeanbarrossilva.mastodonte.app.navigation.navigator.provider
 
 import androidx.fragment.app.Fragment
 import com.jeanbarrossilva.mastodonte.app.R
-import com.jeanbarrossilva.mastodonte.app.navigation.BottomNavigationItemNavigator
+import com.jeanbarrossilva.mastodonte.app.navigation.navigator.BottomNavigationItemNavigator
 import com.jeanbarrossilva.mastodonte.core.profile.Profile
 import com.jeanbarrossilva.mastodonte.core.sample.profile.sample
 import com.jeanbarrossilva.mastodonte.feature.profiledetails.ProfileDetailsFragment
